pyproject.toml
[project]
name = "discord-bot-modern"
version = "2.0.0"
description = "Bot de Discord moderno con Python + Rust, métricas avanzadas y uv"
authors = [
    {name = "Ponce1969", email = "gompatri@gmail.com"}
]
requires-python = ">=3.10"

dependencies = [
    # Discord - Versión exacta que tienes
    "discord.py>=2.4.0",
    
    # Database
    "sqlalchemy>=2.0.31",
    "psycopg2-binary>=2.9.9",
    
    # Environment & Config
    "python-dotenv>=1.0.1",
    
    # Audio & Crypto
    "PyNaCl>=1.5.0",
    
    # HTTP & Async
    "aiohttp>=3.9.5",
    "aiosignal>=1.3.1",
    "async-timeout>=4.0.3",
    
    # Data Processing
    "pandas>=2.2.3",
    "dask[dataframe]>=2024.9.0",
    
    # AI & APIs
    "google-generativeai>=0.7.2",
    "groq>=0.9.0",
    "langchain-community>=0.2.11",
    "openai>=1.0.0",
    
    # Google APIs
    "google-api-python-client>=2.138.0",
    
    # Translation & Text
    "translate>=3.6.1",
    "fuzzywuzzy>=0.18.0",
    "python-levenshtein>=0.26.0",
    
    # Utilities
    "pytz>=2024.2",
    "table2ascii==1.1.3",
    "wcwidth==0.2.13", 
    "tabulate>=0.9.0",
    "Pillow>=10.4.0",
    
    # Low-level dependencies (las que ya tienes)
    "typing_extensions>=4.12.2",
    "yarl>=1.9.4",
    "pycparser>=2.22",
    "multidict>=6.0.5",
    "idna>=3.7",
    "greenlet>=3.0.3",
    "cffi>=1.16.0",
    "attrs>=23.2.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

# Configuración para ruff (herramienta externa compatible con uv)
[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]